Stop bundling the Firebase service account into build output
serviceAccount was a static import from firebase.json, so Next.js
inlined the private key into .next/ build output. Replace it with
web/src/lib/firebase/admin.ts, which lazily initializes firebase-admin
from the FIREBASE_SERVICE_ACCOUNT_B64 env var (base64 JSON) inside
adminAuth()/adminAppCheck(), never at module scope, so npm run build
succeeds without the secret present. actions/auth.ts now goes through
these accessors instead of firebase-admin directly.
