+++ web/README.md
@@ -1,34 +1,51 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+Refitted's admin web app: a Next.js App Router site using React server
+components/actions for admin user management, backed by Firebase Auth
+(session cookies) and Firebase App Check.
 
 ## Getting Started
 
-First, run the development server:
-
 ```bash
+npm ci
 npm run dev
-# or
-yarn dev
 ```
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+Open [http://localhost:3000/admin](http://localhost:3000/admin) with your browser.
 
-You can start editing the page by modifying `pages/index.tsx`. The page auto-updates as you edit the file.
+## Local configuration
 
-[API routes](https://nextjs.org/docs/api-routes/introduction) can be accessed on [http://localhost:3000/api/hello](http://localhost:3000/api/hello). This endpoint can be edited in `pages/api/hello.ts`.
+No config values are committed to this repo — public or not. Create
+`.env.local` (gitignored) with the following:
 
-The `pages/api` directory is mapped to `/api/*`. Files in this directory are treated as [API routes](https://nextjs.org/docs/api-routes/introduction) instead of React pages.
+Build-time, inlined into client code by Next.js (`NEXT_PUBLIC_*` vars are
+public identifiers — Firebase web config and the reCAPTCHA v3 site key —
+but are still kept out of the repo per policy):
 
-## Learn More
+- `NEXT_PUBLIC_FIREBASE_API_KEY`
+- `NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN`
+- `NEXT_PUBLIC_FIREBASE_DATABASE_URL`
+- `NEXT_PUBLIC_FIREBASE_PROJECT_ID`
+- `NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET`
+- `NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID`
+- `NEXT_PUBLIC_FIREBASE_APP_ID`
+- `NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID`
+- `NEXT_PUBLIC_RECAPTCHA_SITE_KEY`
+- `NEXT_PUBLIC_DEV_TOOLS_ENABLED` — set to `true` to enable the Firebase
+  App Check debug token locally.
 
-To learn more about Next.js, take a look at the following resources:
+Runtime, server-only secret:
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+- `FIREBASE_SERVICE_ACCOUNT_B64` — base64 of the Firebase Admin SDK
+  service account JSON (`base64 -w0 firebase.json`). Read lazily in
+  `src/lib/firebase/admin.ts`, so `npm run build` succeeds without it set.
 
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
+Ask a maintainer for the real values. In CI, each of these is set from a
+same-named GitHub secret (see `.github/workflows/build.yml`).
 
-## Deploy on Vercel
+## Architecture
 
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
-
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+Server components and server actions run in trusted server code (deployed
+as a Lambda container via Lambda Web Adapter, built by CDK from another
+branch). The browser never talks to AWS directly. `proxy.ts` runs on the
+Node.js runtime and verifies the session cookie with `firebase-admin` for
+routing; every server action independently re-verifies the session via
+`getAuthenticatedAuth()` as defense in depth.
